Create layout for search bar and profile area
I spent a long time figuring out how to align the right-side of this section properly. I then used placed-content: center and discovered that justify-content was what I needed. Had I not used such css trick, I would create a flex and a grid wrapper. It took a while because I wanted the elements to stay aligned when zoomed out, which was made possible through justify-content